Add ability to modify the image editor toolbar
